qa_doc_agentic_workflow/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.3.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.3.1
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1762421637624/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distro==1.9.0
durationpy==0.10
emoji==2.15.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
faiss-cpu==1.13.1
fastapi==0.124.4
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_ipykernel_1761567969/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1765375773/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kubernetes==34.1.0
langchain==1.1.3
langchain-chroma==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-openai==1.1.3
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lxml==6.0.2
markdown-it-py==4.0.0
marshmallow==3.26.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.23.2
openai==2.11.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
posthog==5.4.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
protobuf==6.33.2
psutil @ file:///opt/miniconda3/conda-bld/psutil_1761896322276/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pypdf==6.4.1
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.3
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1762375617494/work
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tornado @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_63s3osc0_1/croot/tornado_1748956943583/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
unstructured==0.18.21
unstructured-client==0.42.4
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
